+++ extension/global.d.ts
@@ -1,1 +1,3 @@
+declare module '*.css' {}
+
 declare const __BROWSER_TARGET__: 'chrome' | 'firefox';